Fix partition extraction of unformatted partitions.
When image customizer does partition extraction, it runs `fsck` on all
the partitions before and after the extraction. This is a sanity check
to ensure the copy was successful.

However, there is a bug where after partition extraction, `fsck` is
mistakenly run on partitions that aren't formatted (i.e. partitions
that were not assigned a filesystem) and `fsck` unsurprisingly fails.
This change fixes this problem and adds a regression test.