[mic][iso] Include locale for the regenerated initrd.
When dracut drops to the emergency shell, the user cannot type anything
in the console window (sometimes).

This problem is because the locale is missing from the regenerated image.

This changes adds the local to all liveos generated initrds. It adds
about 5kb to the size of the initrd image and is very useful when
debugging failurres.