Container incremental builds.
Change the `build-mic-container.sh` script so that it caches its
resources. In particular, ensure that it only downloads `oras` once.
This allows the `build-mic-container.sh` script to be run multiple
times but it won't do anything unless something has been updated. This
is mostly relying on Docker build's incremental build support.