Clarify logs of build host metadata. (#14)
We currently log information about the build host to assist with
debugging. But some customers have confused the logs for metadata about
the OS being customized. So, make it explicit that the logs are for the
build host. Also, add the distro + version of the base image to the
logs.