Fix shrink partitions race condition. (#26)
After changes to a partition table has been made, including resizing a
partition, all the disk's partition nodes under /dev are removed and
then recreated. We need to ensure we wait for this to complete before
trying to shrink the next partition.

Also, remove the call to `partprobe` since calls to `parted` do this
automatically. So, the `partprobe` call is redundant.