VM Tests: Run Image Customizer container. (#21)
Add executing the Image Customizer container in Docker to the basic
test.

The test suite's Makefile will call the toolkit's Makefile to
automatically rebuild the imagecustomizer binary and the container. This
avoids the user needing to provide a URL to the container. And it
ensures the tests are always running against the code within the
checked-out repo.