VM Tests: Connect to VM using SSH.
Add the ability to connect to VM using SSH so that the contents of the
VM can be verified.

This change includes a wrapper around the paramiko library for running
commands over SSH. This wrapper's primary job is to live log stderr and
stdout while also collecting them as strings.