Better handle filesystem optional features.
Filesystems (e.g. ext4, xfs) sometimes receive new features. When these
features are enabled, anything that loads the filesystem (e.g. kernel,
grub, etc.) must support the new feature or it will refuse to load it.

When formatting a new partition, there are a few considerations for if a
filesystem feature should be enabled:

- Does the version of mkfs support that feature?
- Does the build host kernel support that feature?
- Does the target OS support that feature?

This change ensure that all these considerations are handled correctly
for the ext4 and xfs filesystem types.