Add additional partition types. (#54)
Extend the partition `type` field so that it supports other values. e.g.
`root`, `var`, etc. These new values map to well-known partition type
UUIDs. In addition, allow the user to specify a UUID string in the
`type` field in case we don't yet have a friendly name for that UUID.

Also, add check to ensure the ESP is mounted at `/boot/efi`.