Switch from 'parted' to 'sfdisk'.
Use `sfdisk` and `blockdev` instead of 'parted' and `partprobe`.

This is being done for a number of reasons:

1. There is already a dependency on `sfdisk`. So, this will remove a
dependency.

2. The API of `sfdisk` is a little better structured (IMHO).

3. Feature support for some GPT features (e.g. partition types) came to
`sfdisk` way sooner than they did for `parted`. So, we don't need
version workarounds to use slightly older versions of `sfdisk`.

4. Ubuntu has a "fun" patch added to `parted` that adds in a
`udevadm settle` call to all `parted` operations. This can cause
deadlocks when a app uses `flock`.

In addition, this change updates the partition customization tests so
that they also verify the partition sizes.