Rev up minimum required Dracut version for LiveOS ISO
The logic in LiveOS ISO checks the Dracut version installed on the image
to be customized.

If it is below a certain version, it will always disable SELinux since
it knows SELinux is not supported for that version and the image will
not boot.

However, if it is at or higher, it leave SELinux as-is.

The problem is that this version is not set at 102-8, which has a known
bug blocking booting LiveOS ISO images when SELinux is enabled.

This change updates the version to a new Dracut package that has the
fix necessary for booting LiveOS ISO image with SELinux enabled.