Add multi-kernel support. (#109)
Remove the arbitary restrictions to allow the kernel args to be updated
when multiple kernels are installed.

In addition, remove the logic that sets `"GRUB_DISABLE_RECOVERY"` when
verity is enabled, since the logic can now handle multiple boot entries
in the grub.cfg file when a single kernel is installed.

Also, remove the requirement for `$kernelopts` to exist in the grub.cfg
file, since this logic just complicates the code.