Fix extra command line args in UKIs. (#126)
When generating a UKI, the extra commnand line args are missing from the
UKI. This was caused by the change (#109) that stopped setting
`GRUB_DISABLE_RECOVERY` in the `/etc/default/grub` file. That change
means there are now "recovery mode" entries for each kernel in the
`grub.cfg` file. So, when extracting the kernel args for UKI, we need to
ensure we ignore the "recovery mode" entries.

In addition, add a test that verifies the behavior of `/usr` verity +
UKI. This includes verifying the extra command-line args are present in
the UKI.

Also, update the existing verity tests so that they also check that the
extra command-line args are present (in the grub.cfg file).