Skip UKI test if the 'ukify' command is not available. (#129)
The `ukify` command is relatively new. So, it isn't available except on
fairly recent ditro versions. It isn't available in either Ubuntu 22.04
or AZL2, which is what our build/test pipelines use. So, skip the UKI
test if the `ukify` command isn't available.

Also, fix up the requirements section of the Quick Start guide.

Also, add the `ukify` command to the container.