Remove shrink verity hash partition check.
In an older version of the API, a user would be forced to specify a
dummy filesystem type for the verity hash partition. This filesystem
would be overridden when the verity hash tree was calculated. However,
not only is the dummy filesystem type not required anymore, it is
impossible for a user to specify one. Therefore, we no longer need the
check that ensures we don't try to shrink the verity hash partition.