Cache fstab entries for image.
This change refactors some of the partitions helper functions. The main
goal is to have a cache of the all the partitions' fstab entries.
However, this change also takes the opportunity to do a little bit of
cleanup.

This change is being done in preparation of the change to support
enabling verity of pre-created partitions.