Cache fstab entries for image. (#135)
This change refactors some of the partitions helper functions. The main
goal is to have a cache of the all the partitions' fstab entries.
However, this change also takes the opportunity to do a little bit of
cleanup.

This change is being done in preparation of the change to support
enabling verity of pre-created partitions.

---

### **Checklist**

- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines