Ensure partition metadata is populate in kernel.
The `udevadm settle` command is commonly used to wait for udev to read
a disk and populate the disk's metadata in the kernel. While this works
most of the time, it is well known `udevadm settle` isn't quite enough
to guarantee that the metadata population has finished. In theory,
`udevadm wait` is new command that does provide that guarantee. But that
command isn't widely available yet.

This change works around this problem by adding a wait loop that checks
the kernel's cached disk's metadata and compares it to the expected
values read directly from disk.

This change also replaces the usage of `partprobe` with a call to the
`BLKRRPART` (block device read partition) IOCTL syscall. This was done
because `partprobe` manually handles the partition node updates.
Whereas, calling `BLKRRPART` allows the kernel to handle the update,
which is more efficient and reliable.