2-phase verity. (#141)
Add support for enabling verity in an image in 2 phases:

1. Run MIC/Prism to create the partitions layout.
2. Run MIC/Prism again to enable verity.

This will be useful for scenarios where a base image is created that
contains a set of common configurations and then that base image is
recustomized into multiple variant images.

This change also includes a functional test for this feature.