Add input.image.path (#162)
This PR adds `input.image.path` to the configuration schema as an
optional parameter. It serves as the default value for `--image-file` on
the command-line. While `--image-file` is now optional, one of
`input.image.path` or `--image-file` is still required.

It also cleans up some aspects of the PRs that added
`output.image.path`.