Switch from `parted` to `sfdisk`. (#164)
Switch from using `parted` for creating partitions to using `sfdisk`
instead. The `sfdisk` API is slightly nicer to use. For example, we only
need a single call to `sfdisk` to provision a partition, instead of
multiple calls to `parted`. Also, `sfdisk` acquired some (GPT) features
a lot earlier than `parted`. So, using `sfdisk` allows us to avoid
disabling features for older versions of `parted`. In adddition, this
avoids an annoying patch[1] in Ubuntu which calls `udevadm settle` in
`parted` which can cause our code to deadlock in Ubuntu 24.04.

[1]
https://git.launchpad.net/ubuntu/+source/parted/tree/debian/patches/udevadm-settle.patch