Switch to Kong CLI parser.
Switch from the kingpin CLI parser to using kong instead. Kong is the
successor of kingpin and is more actively maintained. This means that
some of the warts have been fixed and there are more features
available. In particular, kong has an inbuilt option to print usage on
error.

This change also introduces a `build` sub-command. If no command is
specified, `build` is the default command. This will allow us to easily
add other sub-commands in the future. For example, `validate-config`.