Check for input image file existence during config validation stage (#186)
This PR adds file existence checks for the new input.image.path config
property during the config validation stage, for better organization of
code. A similar check is made for output.image.path, where the file may
already exist (it will be truncated), but it must not be a directory.