Install packages in single batch.
When installing packages, instead of calling `tdnf` for each package,
install all the packages in a single call to `tdnf`. This has a number
of advantages including reducing the number of calls to `dracut` and
allowing tdnf's RPM dependency solver to work more effectively.

In addition, add more of the tdnf logs to the DEBUG level to make it
easier for the user to track package installation progress.