MIC: De-duplicate files in LiveOS ISOs. (#215)
- Smaller initrd
  - Stop embedding kernel+shim+bootloader into initrd
- This was a work around to be able to use isomaker as-is. The fix is
stop using isomaker and create the boot image and call mkisofs directly.
- Smaller rootfs
  - Remove the re-created initrd from the squashfs.
- The initrd is always generated under chroot and is created within the
customized rootfs. It was then copied to the iso media. The fix is to
delete (move) it from the customized rootfs as it is not needed to be
there.
  - Remove the boot folder entirely.
- When the input image with all its partitions is mounted, we ended up
copying all the contents including the files at /boot. However, those
files are already on the iso media and do not need to remain in the
squashfs image. The fix is to delete that folder entirely.

The size gains are:
|            | old   | new   |                                    |
|------------|-------|-------|------------------------------------|
| initrd.img |  47MB |  32MB | removal of kernel+shim+bootloader  |
| rootfs.img | 173MB | 138MB | removal of /boot and initrd.img    |
| Test iso   | 250MB | 200MB |                                    |

<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

---

### **Checklist**
- [ ] Tests added/updated
- [ ] Documentation updated (if needed)
- [ ] Code conforms to style guidelines