Slight refactor of verity logic.
This is a setup change for the upcomming 'reinitialize verity' change.

Firstly, break up the verity format helper function
(`customizeVerityImageHelper`) into smaller functions.
This makes the code a little cleaner. Also, the smaller function will
be used in the 'reinitialize verity' change.

Secondly, change the `verityMetadata` variable to be a list everywhere,
instead of a map in some places and a list in other places. Also,
idetify the verity devices by name instead of mount path.