MIC: Factor code out into `IsoArtifactsStore` (#225)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

This change factors out the part that builds up a 'store' of files and
properties from the input image - the `IsoArtifactsStore`.

The artifacts store can be built from either a full image (qcow,
vhd(x)), or a LiveOS ISO.

Then, the store is a unified input from then on to subsequent stages of
creating the final iso.

This way, should we support a new input format, we just need to write a
new converter from that format to the artifact store.

Also this changes moves a couple of function to other parts of the code
where they better match the functionality.

---

### **Checklist**
- [n/a] Tests added/updated
- [n/a] Documentation updated (if needed)
- [ok] Code conforms to style guidelines