Don't copy file ACLs.
Linux has the capability to set extended file permissions (ACLs) to
handle cases where the simple user:group:world permissions model isn't
sufficient. When the `cp` command is set to preserve permissions, it
will also try to preserve the ACLs. However, if the destination
filesystem doesn't support ACLs, it will error out.

This change fixes the problem so that only basic permissions are
preserved, instead of basic and ACLs. Unfortunately, `cp` doesn't have
an option to only preserve basic permissions, it's either all or
nothing. So, this change replaces the use of `cp` with direct I/O calls
within Go.