Fix for when verity hash partition is referenced by UUID.
Ensure the hash partition's UUID is refreshed and reread after
`veritysetup format` is called. This ensures that if the partition is
identified by UUID, then the UUID will be correct in kernel args.

Also, for a regression test, change the existing usr-verity test so
that the verity partitions are referenced by UUID. The root-verity test
will continue to handle the PARTUUID case.