Patch bootctl to not override shim by default. (#247)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

Backup the original shim binary before it gets overwritten by bootctl.
This ensures that the Secure Boot chain is preserved. The reason we do
not disassemble the bootctl and do the clis manually is that the bootctl
is not only doing the folder creation or files copy, but it also
configures the `/boot/efi/loader`, there are some setups under it, for
example:

```
root [ /home/test ]# ls /boot/efi/loader
entries  entries.srel  loader.conf  random-seed
```

---

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines

---

Test pipeline run:
https://dev.azure.com/mariner-org/mariner/_build/results?buildId=819585&view=results