Add initial OTEL instrumentation logic in imagecustomizer (#241)
This PR introduces support for using the OTEL Go sdk to instrument the
tool using OTEL standards and practices. It provides initialization
logic needed for a global tracer to be used across the tool's execution.
For starters, the PR adds a customization time span, with more telemetry
to be added in follow up PRs that help with data collection tied to the
tool's KPIs.

The PR also adds logic to disable telemetry, and uses a no-op tracer
when `--disable-telemetry` is passed in. This follows OTEL standardized
method of disabling telemetry collection in a clean manner.

As part of telemetry initialization, the `telemetry.go` file has been
added to set up the OTEL tracer and the OTLP grpc exporter. The exporter
will send data to a Python service listening at the local port, which
will then forward the data to Azure Monitor.

In the effort of splitting up PRs in chunks, the PR for the complement
Python telemetry service is here:
https://github.com/microsoft/azure-linux-image-tools/pull/249.

---

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines