Support UKI enabled base image customization. (#265)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

In `prepareUki()`, when `copyUkiFiles()` is called, it attempts to find
the `kernel` and its `initramfs` under the boot partition. Once found,
they are placed under `UkiBuildDir` for later use when building the
`UKI` with `ukify`. Additionally, `getKernelToInitramfsMap()` also
heavily depends on the `kernel` and `initramfs` being present in the
boot partition. This function is invoked twice: during both
`prepareUki()` and `createUki()`. Therefore, `cleanupBootPartition()`
should not have to be released into main yet, as we are still in the
process of making the boot partition optional.

Another issue blocking main from customizing a UKI-enabled base image is
the ability to read the latest kernel arguments from an existing UKI. I
have extended the logic - following the structure of
`extractKernelCmdline()` in `partitionutils.go` - to make UKI
customization more robust, enabling `kernel-to-args` mapping to be
extracted from either `grub.cfg` or an existing UKI.

---

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines