MIC: LiveOS: Skip generating the ISO boot image if it is not needed (#271)
When creating a bootable ISO media, the bootloader files (shim and grub)
must be packaged into an image (/boot/grub2/efiboot.img) and specified
to the mkisofs tool. This is required to enable a bootable ISO media.

However, this is not needed for the PXE scenario since the bootloader is
found/obtained by a different mechanism.

This change updates the Image Customizer tool to only generate and
include the efiboot.img if an iso will be generated (i.e. if the output
format is iso - or the output format is pxe and bootstrap initramfs is
specified).

<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

---

### **Checklist**
- [ ] Tests added/updated
- [ ] Documentation updated (if needed)
- [ ] Code conforms to style guidelines