Verity hash signature I/O feature support. (#270)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

This PR adds support for Verity hash signature I/O in Image Customizer,
prepare the appropriate kernel command line arguments (such as
`root-hash-signature=...`), and integrate with the supporting `dracut`
module for boot-time verification.

It also resolves two related bugs. First, it ensures that if
`hashSignaturePath` is not specified for a Verity device, no signature
artifacts are generated or emitted, even if verity-hash is included in
`.output.artifacts`. Second, it fixes the parsing logic for
`systemd.verity_*_options` kernel arguments to properly handle
`root-hash-signature=...` in addition to corruption handling options,
ensuring the signature path is correctly extracted and reapplied when
rebuilding images.

---

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines