Ensure new verity and re-init verity don't conflict.
Reinitialize verity and 2-stage verity can be used at the same time
but we don't currently handle the case where they are both used on the
same partitions. This change detects when this happens and reports an
error.

In the future, we could consider allowing some form of verity settings
override functionality. But we can consider that feature if and when it
is requested by a user.