Add functional tests to workflows.
Run the functional tests as part of the mainline builds and release
builds. Also, allow the functional tests to be optionally run in the
dev workflow but don't enable it by default since it takes a long time
(1 hour) to run.

In addition, fix a minor issue regarding relative paths in a couple of
the tests.