Enable Image Customizer Error Telemetry  (#299)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

Enable error telemetry for imagecustomizer as well as flushing all
telemetry before shutdown.

Needed to add event logic to telemetry hopper because error spans are
structured and processed a bit differently.

For errors, we are keeping a generic error message in the span for now,
to prevent any accidental collection of user data through returned error
strings. In a follow-up effort, we'll have a structured Error Type that
would be included in the span's error field.

---

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines