Image Customizer: clean the /boot folder selectively if kdumpBootFiles=keep is specified (#296)
Clean the /boot folder selectively if kdumpBootFiles=keep is specified.

If set to keep, the Image Customizer tool will not delete any kdump
files found under the boot folder on the full OS image. The kdump files
include:

- a crashdump initramfs image named
`initramfs-<kernel-version>kdump.img`.
- a kernel named `vmlinuz-<kernel-version>` - where its version matches
that of the `initramfs-<kernel-version>kdump.img`.

The default is `none`.

Note that by default, the Image Customizer tool removes the `/boot`
folder from the full OS image. This is because all of its contents have
already been copied unto the root of the PXE artifacts folder.

- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines