Image Customizer: Iso: re-populate the /boot folder without the --no-clobber option
We are disabling the --no-clobber option since the target boot
folder may have common files already from the previous steps.
This can happen when kdump boot files are set to 'keep' in the
input iso which will result in two copies of the kernel vmlinux
file:
- one from the iso media
- one from the full OS image (next to the initramfs kdump.img file).
When the OS is being re-constructed, the attempt to re-copy the
kernel vmlinuz file will fail if we use --no-clobber. So, we disable
it here while copying the boot files from the artifacts.