VMTests suite: Use libvirt API to read console logs.
Currently, the console logs are written directly to file by QEMU.
However, this can cause permission issues, particulartly on hosts with
SELinux enabled (e.g. Fedora).

This change instead reads the console logs using the libvirt API and
then uses Python I/O to write the logs to disk. This avoids the
permission issues since the test process is the one writing the logs,
instead of the more restricted QEMU process.

Similar to the other uses of the libvirt API in the VMTests suite, most
of the code was copy/pasted from the
[LISA](https://github.com/microsoft/lisa/tree/main/lisa/sut_orchestrator/libvirt)
project.