VMTests: Fix when run as 'root' user.
The VMTests suite uses the name of the current user as the name of the
test user it deploys to the OS image. This makes it easier for the user
to manually SSH into the VM when debugging. However, this doesn't work
when the test is being run as "root" (e.g. sudo) since the SSH login is
typically disabled for the "root" user. This change adds code to detect
this and use a different username.