Image Customizer: Iso: recreate image with a single ext4 partition (#333)
Image Customizer: Iso: recreate image with a single ext4 partition

When we are starting from an input iso, we re-construct an image with a
boot fat32 partition and a rootfs ext4 partition. Then, we copy
everything unto that image. This image is later used for OS
customizations (before it is turned back into an iso).

The problem is that the boot fat32 partition cannot preserve all the
file attributes coming from the source image. As a result, the copy
command fails when it's set to preserve attributes. This problem shows
only on Azure Linux 3. It does not repro on Ubuntu 22+.

There is no real need to create the boot partition at all in this flow.
So, this fix drops the boot partition from the temporary image we
create.

- [n/a] Tests added/updated
- [n/a] Documentation updated (if needed)
- [x] Code conforms to style guidelines