VMTests: Fix legacy boot tests. (#346)
In newer versions of libvirt, specifying the `nvram` element in the
domain XML will result in the VM being interpreted as UEFI firmware
based. Whereas in older versions of libvirt, the `nvram` element is
ignored if the other UEFI elements (e.g. `loader`) are not specified.

This change drops the `nvram` element when legacy boot is specified.
This fixes the `libvirt: Domain Config error : no loader path specified
and firmware auto selection disabled` error.