Fix GitHub pages publishing. (#342)
It turns out GitHub workflows won't trigger based on the actions of
other workflows. In this particular case, the release workflow pushing
to the `stable` branch doesn't trigger the GitHub Pages publishing even
though if a human pushed to that branch, it would run.

This change removes the dedicated GitHub Pages publishing workflow and
instead integrates the GitHub Pages publishing directly into the Release
workflows.