Add `reinitializeVerity` API. (#351)
Add API that controls how verity partitions in the base image are
handled. Specifically, whether or not the verity data partitions are set
to read-only or read-write during OS customization.

The primary scenario being targeted is recustomizing an image with a usr
verity partition + UKIs without changing/invalidating the /usr partition
or the UKIs. A functional test has been added for this scenario.