Set container ENTRYPOINT.
Set the `imagecustomizer` executable as the default ENTRYPOINT for the
container.

Technically, the telemetry hopper is already the entrypoint. So, this
change is actually just having the telemetry hopper call
`imagecustomizer`, instead of requiring the user to pass in the command
name. But when the telemetry hopper is removed, `imagecustomizer` will
become the entrypoint.

Also, update the `run.sh` script to include a copy of the telemetry
hopper logic. Now, to use the `run.sh` script, the user must pass
`--entrypoint /usr/local/bin/run.sh` to docker to override the default
command.