Fix overlay SELinux equivalency rules add.
The Azure Linux 3.0's selinux-policy package has a bug where the
SELinux policy is not being rebuilt after the RPM is installed.
This causes problems for Image Customizer's overlays API, since the
`semanage fcontext -a` call requires the policy file to be built.

This change adds a workaround to Image Customizer to detect this case
and triggers an SELinux rebuild when needed.