Add ignoreOverlays param to skip overlay for COSI (#373)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

For COSI generation, connectToExistingImage is called after
customization is done, so when there is overlay in the config,
fstabEntriesToMountPoints() encounters an overlay entry with source
"overlay", isn't in the fstab source types. And overlay should be
excluded from partition metadata so this PR addresses this bug by adding
ignoreOverlays param

---

### **Checklist**
- [ ] Tests added/updated
- [ ] Documentation updated (if needed)
- [ ] Code conforms to style guidelines

Signed-off-by: Elaine Zhao <elainezhao@microsoft.com>