Workaround 'sudo' problem in GitHub workflows.
The GitHub workflows are experiencing the error: "sudo: unable to
resolve host 8e70e4e9c000000: Name or service not known" when the
`sudo` command is used. This is polluting the stderr, which can cause
problems when parsing output of `docker image load` command. Workaround
this problem by making the output parsing more robust.