Image Customizer: assign a randomly generated label to the ISO media (#400)
Image Customizer: assign a randomly generated label to the ISO media

When IC builds an ISO, it configures the boot process to look for a
device with a file system labeled 'CDROM'. This is used to find the
kernel/initramfs and also to find the root squashfs image.

We have run into a case where a target baremetal machine had a partition
labeled CDROM (for some reason!). As a result, the boot process was not
able to find the above artifacts and the boot failed.

This fix ensures that the media label is unique, and ensures that its
references are updated to use that same unique value.

The unique value is a uuid.

Checklist:
- [ ] Tests added/updated
- [ ] Documentation updated (if needed)
- [x] Code conforms to style guidelines