Image Customizer: rename the module from azurelinux to azure-linux-image-tools (#407)
Image Customizer: rename the module from azurelinux to
azure-linux-image-tools

In order for other golang tools to be able to import the packages
implemented in the `azure-linux-image-tools` repo, we need to ensure the
module name/path matches its github path in `go.mod`. Along with that,
all other referencing code must be updated to match this module new
name.

### **Checklist**
- [ ] Tests added/updated
- [ ] Documentation updated (if needed)
- [x] Code conforms to style guidelines