Improve error messages. (#409)
Improve the error message when user passes a nonsense file as the base
image. Previously, the error was "failed to find rootfs partition". Now,
it is "image does not contain a partition table".

Also, add some high-level errors for these operations:

- Partition customization
- Reset partition UUIDs
- OS customization
- Verity provisioning
- Creating UKIs
- Outputting artifacts

These extra errors will make it clearer where in the customization
process the error occurred.