Support forking before publishing new release.
Add support for forking a release branch before doing a release.

Firstly, update the `Release (major/minor)` workflow with an option to
not publish the release and only fork the release branch.

Secondly, update the `Release (patch)` workflow's logic that increments
the patch version so that it supports the case where no release has
occured for the major/minor version yet.