For 'inject-files', shrink non-verity partitions for COSI output. (#419)
When using the 'inject-files' subcommand, if the output format is COSI
then shrink the non-verity partitions. This is to be more consistent
with the customize subcommand, which shrinks all partitions when
outputting as COSI.

It is not possible for 'inject-files' to shrink verity partitions since
that would invalidate the verity hash tree. In the future, we may add an
API to allow the customize command to shrink verity partitions even when
not outputting as COSI.