Fix UKI root verity functional test failure (#438)
<!-- Description: Please provide a summary of the changes and the
motivation behind them. -->

For root verity, we cannot make the partition read-only during
customization because the resolve config requires write access.
Additionally, for the second customization in this functional test case,
the UKI won't be carried over, so we need to add extra fields to the MIC
config.

---

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines

Test:
https://github.com/microsoft/azure-linux-image-tools/actions/runs/17870292411/job/50822285967
See `--- PASS: TestCustomizeImageVerityRootUki (54.22s)`