Detect GRUB + xfs + sparse file problem.
GRUB is unable to read sparse files on an xfs filesystem. So, add logic
that fails the build if a file GRUB care about (e.g. kernel or
initramfs) is on an XFS filesystem and is sparse.

While this doesn't prevent the underlying problem causing the files to
be sparse in the first place, it ensures that the problem results in a
build-time error instead of a runtime error.