Remove temp resources between tests.
Currently, the test temporary directory is removed at the end of all the
tests. So, as more tests are added, the temp directory has gotten larger
and larger and now the tests are failing due to out of disk space
errors.

This change removes the test specific temporary directories after a test
has completed. This change also makes the creation of the temporary
directories more consistent between tests.