Remove dead code.
This change removes the `imager` and `isomaker` tools. Then it deletes a
large number of dead functions, which were found using the `deadcode` Go
tool.

There is more code that can be removed from functions that are still
used but have unused functionality. But that will be a more complex
change.