Add merge queue trigger to workflow.
Trigger the main workflow on merge_queue event. This will cause the
workflow to run when a change is placed in a merge queue.

The purpose of this change is to ensure the tests suite succeed before
a change is merged into main, but avoid running these long tests on
every iteration of a PR.