Fix when /boot is on an XFS filesystem. (#451)
GRUB 2.06, which is used in AZL3, doesn't support the xfs feature
`nrext64`. However, currently MIC's default features for AZL3 include
the `nrext64` feature. So, if the `/boot` directory is on an xfs
partition, then the image will fail to boot. This change fixes this by
excluding the `nrext64` feature when the `/boot` directory is on an xfs
partition.