Create CLI options struct. (#459)
Whenever a new CLI arg is added, currently 50+ tests have to be updated
to handle the new param in the `CustomizeImageWithConfigFile` function.
This can be pretty annoying.

This change creates a new `ImageCustomizerOptions` struct to store all
the options provided on the command-line. Now, the the CLI options can
be passed by struct instead of new params being added each time.

The existing functions `CustomizeImageWithConfigFile` and
`CustomizeImage` have not been changed to avoid needing to update all
the existing tests. Also, using the options struct can be quite verbose.