Update quick start guide by specifying Linux distribution and version in Image Creator command. (#464)
The quick-start guide documentation has been updated to clearly specify
the Linux distribution and version parameters in the Image Creator
command example. This helps users by indicating the exact distribution
(azurelinux) and version (3.0) to use.

Included the --distro azurelinux parameter in the command example
Included the --distro-version 3.0 parameter in the command example

### **Checklist**
- [x] Tests added/updated
- [x] Documentation updated (if needed)
- [x] Code conforms to style guidelines